feat: render system implements flip and provisional rotation - Updated MovementSystem, RenderSystem, SpriteComponent, ColliderComponent - Adds flip property to SpriteComponent - MovementSystem calculates orientation based on velocity and determines flip on collisions - RenderSystem handles flip/rotation - ColliderComponent adds bounce detection

+++ src/main/java/com/mycompany/animacionpecera/VelocityComponent.java
@@ -10,10 +10,10 @@ package com.mycompany.animacionpecera;
  */
 public class VelocityComponent extends Component{
     
-    public float velX;
-    public float velY;
+    public double velX;
+    public double velY;
     
-    public VelocityComponent(float velX, float velY) {
+    public VelocityComponent(double velX, double velY) {
         this.velX = velX;
         this.velY = velY;
     }
